Switched to a "layout" based-generation and created first plate model
Generating cases from the plate dxf, as I was doing before, has proven
too unpredictable. Because the total size or layout of the plate could not be
known, too many concessions needed to be made to make a functional case,
such that the case could not be generated purely from the dxf.

Because of this, I've instead made the decision to generate the case
from a text layout. Eventually, a script should accompany the scad
file(s) to convert from a keyboard-layout-editor.com export and generate
a case from that.

The old dxf-based files have been tagged as v1.0.0, and can be accessed
at that tag if needed.

This first "v2" commit contains a scad file for generating a "reinforced
plate based on a two-dimensional vector layout. In this first version,
only 1u keys are supported.