Calculate cumulative distance in plate.scad
Added a 2D array, derived from the layout, where each element represents
the cumulative distance (in cherry units) of the right edge of the
corresponding element in the layout. For example,
	[
		[1,1,1,1],
		[1,2,3,4]
	]
beomes
	[
		[1,2,3,4],
		[1,3,6,10]
	]

A series of commented-out functions showing how the function to dervive
this array was written have been left in this commit for future
reference, but will be removed in the next commit.