Remove cantilevers & add holes for screws & inserts
NOTE:
This commit (and likely several commits following this one) contains
code that is specific to the current layout (in this case, the placement
of the screws/inserts). I want to make of working version of this
iteration of the keyboard case so that I can start using it and assess
any strengths and weaknesses before moving to the next iteration
(generation from keyboard-layout-editor.com JSON). I may generalize it
later, if I don't decide to do a complete rewrite to accomodate new
features. In the meantime, it should be fairly straightforward to modify
for another layout.