Parametrically determined connection holes
Currently "starts" the holes at the back wall of the case. Can be a
problem if they are too tall to "fit" on the back wall, causing the
forward slope to be built in front of them. A few possible options exist
which I will explore. For now, just making sure the back wall is tall
enough to accomodate the connection holes will suffice.